# Reduce memory use, improve pixel classification measurement performance