--- a/qupath-core-processing/src/main/java/qupath/opencv/tools/OpenCVTools.java
+++ b/qupath-core-processing/src/main/java/qupath/opencv/tools/OpenCVTools.java
@@ -422,8 +422,10 @@ public class OpenCVTools {
 			} else {
 				opencv_core.merge(new MatVector(channels.toArray(Mat[]::new)), dest);
 			}
-//			scope.deallocate();
 		}
+
+		if (logger.isTraceEnabled())
+			logger.trace("Merged channels: {}", Arrays.stream(dest.createIndexer().sizes()).mapToObj(l -> l).toList());
 
 		return dest;
 	}