--- a/qupath-gui-fx/src/main/java/qupath/lib/gui/panes/SelectedMeasurementTableView.java
+++ b/qupath-gui-fx/src/main/java/qupath/lib/gui/panes/SelectedMeasurementTableView.java
@@ -217,19 +217,6 @@ public class SelectedMeasurementTableView implements PathObjectSelectionListener
 			tableMeasurements.refresh();
 			delayedUpdate = false;
 		}
-
-//		// Check if objects are outside hierarchy
-//		if (imageData != null) {
-//			for (PathObject pathObject : getSelectedObjectList()) {
-//				if (PathObjectTools.hierarchyContainsObject(imageData.getHierarchy(), pathObject)) {
-//					tableModel.setImageData(this.imageData, getSelectedObjectList());
-//					tableMeasurements.refresh();
-//					return;
-//				}
-//			}
-//		}
-//		tableModel.setImageData(null, getSelectedObjectList());
-//		tableMeasurements.refresh();
 	}
 
 	@Override