… qupath-core-processing/src/main/java/qupath/opencv/ops/ImageOps.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/ViewerManager.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/tools/handlers/AbstractPathROIToolEventHandler.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/tools/handlers/AbstractPathToolEventHandler.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/tools/handlers/AbstractPolyROIToolEventHandler.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/tools/handlers/BrushToolEventHandler.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/tools/handlers/MoveToolEventHandler.java
… qupath-gui-fx/src/main/java/qupath/lib/gui/viewer/tools/handlers/ToolUtils.java