Numerous changes aiming to reduce memory use, including:
* Preventing objects (and object hierarchies) from being retained for a time after an image has been closed
* Improving the efficiency of `ImageOps` with padding
* Parallelising pixel classification measurements

The last two of these should improve pixel classification generally, and making measurements in particular.

It is the proposed fix for https://github.com/qupath/qupath/issues/1322

I obtained some rough timings using *Run for project* with 3 images and an extreme pixel classifier with 250 output channels:

```groovy
setImageType('BRIGHTFIELD_H_E');
setColorDeconvolutionStains('{"Name" : "H&E default", "Stain 1" : "Hematoxylin", "Values 1" : "0.65111 0.70119 0.29049", "Stain 2" : "Eosin", "Values 2" : "0.2159 0.8012 0.5581", "Background" : " 255 255 255"}');
clearAllObjects();
resetSelection();
//createAnnotationsFromPixelClassifier("My threshold", 0.0, 0.0, "INCLUDE_IGNORED", "SPLIT") // Alternative option
createAnnotationsFromPixelClassifier("My threshold", 0.0, 0.0, "INCLUDE_IGNORED")
selectAnnotations();
addPixelClassifierMeasurements("Busy classifier", "Busy classifier")
println getAnnotationObjects().measurements
```

With larger images / ROIs, the changes can result in huge speed improvements (tested using M1 Max, macOS 13.5.2):

### v0.5.0-SNAPSHOT

```
INFO: Starting script at Fri Sep 15 13:54:32 BST 2023
INFO: [[Busy classifier: Stroma %:57.11750411987305, Busy classifier: Stroma area µm^2:3.7282948E7, Busy classifier: Tumor %:42.88249588012695, Busy classifier: Tumor area µm^2:2.799117E7]]
INFO: Total run time: 14.97 seconds
INFO: Starting script at Fri Sep 15 13:54:47 BST 2023
INFO: [[Busy classifier: Stroma %:24.72903823852539, Busy classifier: Stroma area µm^2:122532.3984375, Busy classifier: Tumor %:75.27095794677734, Busy classifier: Tumor area µm^2:372967.65625]]
INFO: Total run time: 0.42 seconds
INFO: Starting script at Fri Sep 15 13:54:48 BST 2023
INFO: [[Busy classifier: Stroma %:42.06702423095703, Busy classifier: Stroma area µm^2:2.7533646E7, Busy classifier: Tumor %:57.93297576904297, Busy classifier: Tumor area µm^2:3.7918208E7]]
INFO: Total run time: 9.97 seconds
```

## v0.4.4
```
INFO: Starting script at Fri Sep 15 13:57:43 BST 2023
INFO: [[Busy classifier: Stroma %:57.11750411987305, Busy classifier: Stroma area µm^2:3.7282948E7, Busy classifier: Tumor %:42.88249588012695, Busy classifier: Tumor area µm^2:2.799117E7]]
INFO: Total run time: 78.18 seconds
INFO: Starting script at Fri Sep 15 13:59:01 BST 2023
INFO: [[Busy classifier: Stroma %:24.72903823852539, Busy classifier: Stroma area µm^2:122532.3984375, Busy classifier: Tumor %:75.27095794677734, Busy classifier: Tumor area µm^2:372967.65625]]
INFO: Total run time: 0.49 seconds
INFO: Starting script at Fri Sep 15 13:59:02 BST 2023
INFO: [[Busy classifier: Stroma %:42.06702423095703, Busy classifier: Stroma area µm^2:2.7533646E7, Busy classifier: Tumor %:57.93297576904297, Busy classifier: Tumor area µm^2:3.7918208E7]]
INFO: Total run time: 74.90 seconds
```